Binary: extract package first, then verify the files
The goal is to be able to push extraction to a background
process. verify_binary() is not ideal for that, because
it can be relatively expensive, and we cannot perform it
in the same background process as the extraction, unless
we start an new R process.

With the rewrite we verify the already extracted files,
and that is much cheaper, so we can run it in the master
process.